refactor: hydrate records outside query service
