+++ plugins/data-sources/types.d.ts
@@ -47,7 +47,6 @@ export interface IQueryService {
     offset?: number;
     orderBy: string;
     orderDirection: string;
-    columns: Column[];
   }): Promise<[]>;
   getRecordsCount(payload: {
     tableName: string;
@@ -56,7 +55,6 @@ export interface IQueryService {
   getRecord(payload: {
     tableName: string;
     recordId: string;
-    columns: Column[];
   }): Promise<Record<string, unknown> | undefined>;
   updateRecord({
     tableName: string,
